Move json_library hook to Ueberauth module
