+++ lib/ueberauth.ex
@@ -170,6 +170,27 @@ defmodule Ueberauth do
     conn.assigns[:ueberauth_auth]
   end
 
+  @doc """
+  A json library is required for Ueberauth to operate.
+  In config.exs your implicit or expicit configuration is:
+    config ueberauth,
+      json_library: Jason # defaults to Jason but can be configured to Poison
+
+  In mix.exs you will need something like:
+    def deps() do
+      [
+        ...
+        {:jason, :version} # or {:poison, :version}
+      ]
+    end
+
+  This file will serve underlying Ueberauth libraries as a hook to grab the
+  configured json library.
+  """
+  def json_library do
+    Application.get_env(:ueberauth, :json_library, Jason)
+  end
+
   @doc false
   def init(opts \\ []) do
     {provider_list, opts} = Keyword.pop(opts, :providers, :all)
